// apps/api/src/arbitration/entities/abstention-reason.enum.ts
export enum AbstentionReason {
  PARSE_FAILURE = 'PARSE_FAILURE',
  NO_CITED_EVIDENCE = 'NO_CITED_EVIDENCE',
  UNRESOLVED_INTEGRITY_FLAGS = 'UNRESOLVED_INTEGRITY_FLAGS',
  CONTRADICTORY_OR_MISSING_EVIDENCE = 'CONTRADICTORY_OR_MISSING_EVIDENCE',
  LOW_CONFIDENCE = 'LOW_CONFIDENCE',
  PROVIDER_ERROR = 'PROVIDER_ERROR',
}